+++ Code/GraphMol/MolDraw2D/CMakeLists.txt
@@ -42,16 +42,9 @@ rdkit_library(MolDraw2D MolDraw2D.cpp MolDraw2DSVG.cpp
   SubstructMatch Subgraphs GraphMol EigenSolvers )
 target_compile_definitions(MolDraw2D PRIVATE RDKIT_MOLDRAW2D_BUILD)
 
+# if(RDK_BUILD_QT_DEMO && !RDK_BUILD_QT_SUPPORT) error
 if(RDK_BUILD_QT_SUPPORT)
-  target_compile_definitions(MolDraw2D PUBLIC "-DRDK_BUILD_QT_SUPPORT")
-  find_package(Qt5 COMPONENTS Widgets OpenGL REQUIRED)
-  target_sources(MolDraw2D PRIVATE MolDraw2DQt.cpp DrawTextQt.cpp)
-  target_link_libraries(MolDraw2D PRIVATE Qt5::Widgets Qt5::OpenGL)
-  rdkit_headers(MolDraw2DQt.h DrawTextQt.h DEST GraphMol/MolDraw2D)
-  if(RDK_INSTALL_STATIC_LIBS)
-     target_link_libraries(MolDraw2D_static PRIVATE Qt5::Widgets Qt5::OpenGL)
-     target_sources(MolDraw2D_static PRIVATE MolDraw2DQt.cpp DrawTextQt.cpp)
-  endif()
+  add_subdirectory(Qt)
 endif(RDK_BUILD_QT_SUPPORT)
 
 if(RDK_BUILD_CAIRO_SUPPORT)
@@ -100,13 +93,6 @@ if(RDK_BUILD_FREETYPE_SUPPORT)
     endif()
     rdkit_headers(DrawTextFTCairo.h DEST GraphMol/MolDraw2D)
   endif()
-  if(RDK_BUILD_QT_SUPPORT)
-    target_sources(MolDraw2D PRIVATE DrawTextFTQt.cpp)
-    if(RDK_INSTALL_STATIC_LIBS)
-      target_sources(MolDraw2D_static PRIVATE DrawTextFTQt.cpp)
-    endif()
-    rdkit_headers(DrawTextFTQt.h DEST GraphMol/MolDraw2D)
-  endif()
 endif(RDK_BUILD_FREETYPE_SUPPORT)
 
 rdkit_test(moldraw2DTest1 test1.cpp LINK_LIBRARIES
@@ -115,18 +101,9 @@ rdkit_test(moldraw2DTest1 test1.cpp LINK_LIBRARIES
 rdkit_catch_test(moldraw2DTestCatch catch_main.cpp catch_tests.cpp LINK_LIBRARIES
   MolDraw2D CIPLabeler )
 
-if(RDK_BUILD_QT_SUPPORT)
-rdkit_catch_test(moldraw2DTestQt catch_qt.cpp LINK_LIBRARIES
-  MolDraw2D Qt5::Widgets Qt5::OpenGL)
-endif(RDK_BUILD_QT_SUPPORT)
-
 rdkit_test(moldraw2DRxnTest1 rxn_test1.cpp LINK_LIBRARIES
   MolDraw2D )
 
-if(RDK_BUILD_QT_DEMO)
-  add_subdirectory(QTDemo)
-endif(RDK_BUILD_QT_DEMO)
-
-if(RDK_BUILD_PYTHON_WRAPPERS)
+  if(RDK_BUILD_PYTHON_WRAPPERS)
 add_subdirectory(Wrap)
 endif()
