+++ Code/GraphMol/MolDraw2D/Qt/Demo/LICENSE
