+++ Code/GraphMol/MolDraw2D/Qt/Demo/QT4SelectItems.cc
