+++ Code/GraphMol/MolDraw2D/Qt/Demo/QTGet2Strings.H
