+++ Code/GraphMol/MolDraw2D/Qt/Demo/RDKitSVMainWindow.H
