+++ Code/GraphMol/MolDraw2D/Qt/Demo/RDKitSVPanel.cc
