+++ Code/GraphMol/MolDraw2D/Qt/Demo/RDKitSVSettings.cc
