+++ Code/GraphMol/MolDraw2D/Wrap/CMakeLists.txt
@@ -11,6 +11,3 @@ rdkit_python_extension(rdMolDraw2D
                        LINK_LIBRARIES MolDraw2D ${QtDrawingLibs} )
 
 add_pytest(pyMolDraw2D ${CMAKE_CURRENT_SOURCE_DIR}/testMolDraw2D.py)
-if(RDK_BUILD_QT_SUPPORT)
-  add_pytest(pyMolDraw2DQt ${CMAKE_CURRENT_SOURCE_DIR}/testMolDraw2DQt.py)
-endif(RDK_BUILD_QT_SUPPORT)
